ready to push cpp00

+++ cpp00/ex00/Megaphone.cpp
@@ -6,14 +6,14 @@
 /*   By: lchan <lchan@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/12 18:55:19 by lchan             #+#    #+#             */
-/*   Updated: 2022/09/12 19:31:50 by lchan            ###   ########.fr       */
+/*   Updated: 2022/09/16 15:40:06 by lchan            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <iostream>
 #include <cctype>
 #include <stdio.h>
-#include <string>
+#include <string.h>
 
 int	main (int ac, char **av)
 {
@@ -22,7 +22,7 @@ int	main (int ac, char **av)
 	if (ac > 1)
 	{
 		while (av[++tab_index])
-			for(int i = 0; i < strlen(av[tab_index]); i++)
+			for(size_t i = 0; i < strlen(av[tab_index]); i++)
 				putchar(toupper(av[tab_index][i]));
 		std::cout << std::endl;
 	}
